separate entry points for node.js vs browsers
