+++ example-typescript/package.json
@@ -11,7 +11,7 @@
     }
   },
   "dependencies": {
-    "url-metadata": "^5.0.4"
+    "url-metadata": "^5.0.5"
   },
   "devDependencies": {
     "buffer": "^6.0.3",
